(PC-37106)[PRO] chore: move indiv. offer summary stocks screen in appropriate folder (new arch)
