+++ pro/src/commons/utils/getDepartmentCode.ts
