+++ pro/src/components/IndividualOffer/IndividualOfferLayout/OfferPublicationEdition/OfferPublicationEdition.tsx
@@ -5,8 +5,8 @@ import { api } from 'apiClient/api'
 import { GetIndividualOfferWithAddressResponseModel } from 'apiClient/v1'
 import { GET_OFFER_QUERY_KEY } from 'commons/config/swrQueryKeys'
 import { useNotification } from 'commons/hooks/useNotification'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { serializeDateTimeToUTCFromLocalDepartment } from 'components/IndividualOffer/StocksEventEdition/serializers'
-import { getDepartmentCode } from 'components/IndividualOffer/utils/getDepartmentCode'
 import fullEditIcon from 'icons/full-edit.svg'
 import { Button } from 'ui-kit/Button/Button'
 import { ButtonVariant } from 'ui-kit/Button/types'
