+++ pro/src/components/IndividualOffer/StocksEventCreation/StocksCalendar/StocksCalendarForm/StocksCalendarForm.tsx
@@ -4,8 +4,8 @@ import { FormProvider, useForm } from 'react-hook-form'
 import { api } from 'apiClient/api'
 import { GetIndividualOfferWithAddressResponseModel } from 'apiClient/v1'
 import { useNotification } from 'commons/hooks/useNotification'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { MandatoryInfo } from 'components/FormLayout/FormLayoutMandatoryInfo'
-import { getDepartmentCode } from 'components/IndividualOffer/utils/getDepartmentCode'
 import { ScrollToFirstHookFormErrorAfterSubmit } from 'components/ScrollToFirstErrorAfterSubmit/ScrollToFirstErrorAfterSubmit'
 import { DialogBuilder } from 'ui-kit/DialogBuilder/DialogBuilder'
 import { RadioGroup } from 'ui-kit/form/RadioGroup/RadioGroup'
