+++ pro/src/components/IndividualOffer/StocksThing/StocksThing.tsx
@@ -20,6 +20,7 @@ import { isOfferDisabled } from 'commons/core/Offers/utils/isOfferDisabled'
 import { useNotification } from 'commons/hooks/useNotification'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { getToday, getYearMonthDay, isDateValid } from 'commons/utils/date'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { getLocalDepartementDateTimeFromUtc } from 'commons/utils/timezone'
 import { DuoCheckbox } from 'components/DuoCheckbox/DuoCheckbox'
 import { FormLayout } from 'components/FormLayout/FormLayout'
@@ -36,7 +37,6 @@ import { TextInput } from 'ui-kit/form/TextInput/TextInput'
 import { ListIconButton } from 'ui-kit/ListIconButton/ListIconButton'
 
 import { DialogStockThingDeleteConfirm } from '../DialogStockDeleteConfirm/DialogStockThingDeleteConfirm'
-import { getDepartmentCode } from '../utils/getDepartmentCode'
 import { getSuccessMessage } from '../utils/getSuccessMessage'
 
 import { ActivationCodeFormDialog } from './ActivationCodeFormDialog/ActivationCodeFormDialog'
