+++ pro/src/components/IndividualOffer/StocksThing/utils/buildInitialValues.ts
@@ -5,8 +5,8 @@ import {
   GetOfferStockResponseModel,
 } from 'apiClient/v1'
 import { FORMAT_ISO_DATE_ONLY, isDateValid } from 'commons/utils/date'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { getLocalDepartementDateTimeFromUtc } from 'commons/utils/timezone'
-import { getDepartmentCode } from 'components/IndividualOffer/utils/getDepartmentCode'
 
 import { STOCK_THING_FORM_DEFAULT_VALUES } from '../constants'
 import { StockThingFormValues } from '../types'
