+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummary/components/IndividualOfferSummaryScreen.tsx
@@ -12,6 +12,7 @@ import { OFFER_WIZARD_MODE } from 'commons/core/Offers/constants'
 import { getIndividualOfferUrl } from 'commons/core/Offers/utils/getIndividualOfferUrl'
 import { useNotification } from 'commons/hooks/useNotification'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { getOfferConditionalFields } from 'commons/utils/getOfferConditionalFields'
 import { DisplayOfferInAppLink } from 'components/DisplayOfferInAppLink/DisplayOfferInAppLink'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
@@ -28,7 +29,6 @@ import { Callout } from 'ui-kit/Callout/Callout'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
 
 import { serializeDateTimeToUTCFromLocalDepartment } from '../../../../components/IndividualOffer/StocksEventEdition/serializers'
-import { getDepartmentCode } from '../../../../components/IndividualOffer/utils/getDepartmentCode'
 
 import { EventPublicationForm } from './EventPublicationForm/EventPublicationForm'
 import { EventPublicationFormValues } from './EventPublicationForm/types'
