+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummaryStocks/IndividualOfferSummaryStocks.tsx
@@ -4,13 +4,14 @@ import { useIndividualOfferContext } from 'commons/context/IndividualOfferContex
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { useOfferWizardMode } from 'commons/hooks/useOfferWizardMode'
 import { IndividualOfferLayout } from 'components/IndividualOffer/IndividualOfferLayout/IndividualOfferLayout'
-import { StocksCalendarSummaryScreen } from 'components/IndividualOffer/StocksCalendarSummaryScreen/StocksCalendarSummaryScreen'
-import { StocksSummaryScreen } from 'components/IndividualOffer/StocksSummaryScreen/StocksSummaryScreen'
 import { OFFER_WIZARD_STEP_IDS } from 'components/IndividualOfferNavigation/constants'
 import { ActionBar } from 'pages/IndividualOffer/components/ActionBar/ActionBar'
 import { Spinner } from 'ui-kit/Spinner/Spinner'
 
-const StocksSummary = (): JSX.Element | null => {
+import { IndividualOfferSummaryStocksCalendarScreen } from './components/IndividualOfferSummaryStocksCalendarScreen/IndividualOfferSummaryStocksCalendarScreen'
+import { IndividualOfferSummaryStocksScreen } from './components/IndividualOfferSummaryStocksScreen/IndividualOfferSummaryStocksScreen'
+
+const IndividualOfferSummaryStocks = (): JSX.Element | null => {
   const mode = useOfferWizardMode()
   const { offer } = useIndividualOfferContext()
 
@@ -25,9 +26,9 @@ const StocksSummary = (): JSX.Element | null => {
   return (
     <IndividualOfferLayout title="Récapitulatif" offer={offer} mode={mode}>
       {isEventWithOpeningHoursEnabled && offer.isEvent ? (
-        <StocksCalendarSummaryScreen offer={offer} />
+        <IndividualOfferSummaryStocksCalendarScreen offer={offer} />
       ) : (
-        <StocksSummaryScreen />
+        <IndividualOfferSummaryStocksScreen />
       )}
       <ActionBar step={OFFER_WIZARD_STEP_IDS.SUMMARY} isDisabled={false} />
     </IndividualOfferLayout>
@@ -36,4 +37,4 @@ const StocksSummary = (): JSX.Element | null => {
 
 // Below exports are used by react-router
 // ts-unused-exports:disable-next-line
-export const Component = StocksSummary
+export const Component = IndividualOfferSummaryStocks
