+++ pro/src/pages/IndividualOfferSummary/IndividualOfferSummaryStocks/components/IndividualOfferSummaryStocksScreen/IndividualOfferSummaryStocksScreen.module.scss
