+++ pro/src/pages/IndividualOfferSummary/components/StockThingSection/StockThingSection.spec.tsx
