+++ pro/src/pages/IndividualOfferSummary/components/StockThingSection/StockThingSection.tsx
