+++ pro/src/pages/IndividualOffers/IndividualOffersContainer/components/IndividualOffersTable/components/IndividualOfferRow/components/OfferNameCell/OfferNameCell.tsx
@@ -4,9 +4,9 @@ import { Link } from 'react-router'
 import { ListOffersOfferResponseModel } from 'apiClient/v1'
 import { OFFER_STATUS_SOLD_OUT } from 'commons/core/Offers/constants'
 import { FORMAT_DD_MM_YYYY_HH_mm } from 'commons/utils/date'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { pluralize } from 'commons/utils/pluralize'
 import { formatLocalTimeDateString } from 'commons/utils/timezone'
-import { getDepartmentCode } from 'components/IndividualOffer/utils/getDepartmentCode'
 import { getCellsDefinition } from 'components/OffersTable/utils/cellDefinitions'
 import { Tag } from 'design-system/Tag/Tag'
 import fullErrorIcon from 'icons/full-error.svg'
