+++ pro/src/pages/IndividualOffers/IndividualOffersContainer/components/IndividualOffersTable/components/IndividualOfferRow/components/OfferStatusCell/OfferStatusCell.tsx
@@ -4,8 +4,8 @@ import { isAfter } from 'date-fns'
 import { ListOffersOfferResponseModel, OfferStatus } from 'apiClient/v1'
 import { useActiveFeature } from 'commons/hooks/useActiveFeature'
 import { FORMAT_DD_MM_YYYY_HH_mm } from 'commons/utils/date'
+import { getDepartmentCode } from 'commons/utils/getDepartmentCode'
 import { formatLocalTimeDateString } from 'commons/utils/timezone'
-import { getDepartmentCode } from 'components/IndividualOffer/utils/getDepartmentCode'
 import { getCellsDefinition } from 'components/OffersTable/utils/cellDefinitions'
 import { StatusLabel } from 'components/StatusLabel/StatusLabel'
 import { Tag, TagVariant } from 'design-system/Tag/Tag'
